+++ ImageCropperDemo/MainPage.xaml.cs
@@ -1,16 +1,21 @@
-﻿namespace ImageCropperDemo
+﻿using Microsoft.Extensions.Logging;
+
+namespace ImageCropperDemo
 {
     public partial class MainPage : ContentPage
     {
-        public MainPage()
+        private readonly ILogger<MainPage> _logger;
+
+        public MainPage(ILogger<MainPage> logger)
         {
             InitializeComponent();
+            _logger = logger;
         }
 
         protected async void OnClickedRectangle(object sender, EventArgs e)
         {
             imageView.Source = null;
-            new ImageCropper.Maui.ImageCropper()
+            await new ImageCropper.Maui.ImageCropper
             {
                 //                PageTitle = "Test Title",
                 //                AspectRatioX = 1,
@@ -22,8 +27,9 @@
                         imageView.Source = ImageSource.FromFile(imageFile);
                     });
                 },
-                Failure = () => {
-                    Console.WriteLine("Error capturing an image to crop.");
+                Failure = () =>
+                {
+                    _logger.LogWarning("Failed to capture or crop image");
                 }
             }.Show(this);
         }
@@ -31,7 +37,7 @@
         private async void OnClickedCircle(object sender, EventArgs e)
         {
             imageView.Source = null;
-            new ImageCropper.Maui.ImageCropper()
+            await new ImageCropper.Maui.ImageCropper()
             {
                 CropShape = ImageCropper.Maui.ImageCropper.CropShapeType.Oval,
                 Success = (imageFile) =>
@@ -41,8 +47,9 @@
                         imageView.Source = ImageSource.FromFile(imageFile);
                     });
                 },
-                Failure = () => {
-                    Console.WriteLine("Error capturing an image to crop.");
+                Failure = () =>
+                {
+                    _logger.LogWarning("Failed to capture or crop image");
                 }
             }.Show(this);
         }